refactor: modularize parking spot script
